python/requirements.txt
wolframalpha
python-dotenv
pyinstaller
